[chain][network] Factor out block reward from internal chain to proper package
