+++ internal/chain/reward.go
@@ -7,7 +7,6 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/rlp"
 	"github.com/harmony-one/harmony/block"
-	"github.com/harmony-one/harmony/common/denominations"
 	"github.com/harmony-one/harmony/consensus/engine"
 	"github.com/harmony-one/harmony/consensus/reward"
 	"github.com/harmony-one/harmony/consensus/votepower"
@@ -18,76 +17,15 @@ import (
 	"github.com/harmony-one/harmony/numeric"
 	"github.com/harmony-one/harmony/shard"
 	"github.com/harmony-one/harmony/staking/availability"
+	"github.com/harmony-one/harmony/staking/network"
 	"github.com/harmony-one/harmony/staking/slash"
 	"github.com/pkg/errors"
 )
-
-var (
-	// BlockReward is the block reward, to be split evenly among block signers.
-	BlockReward = new(big.Int).Mul(big.NewInt(24), big.NewInt(denominations.One))
-	// BaseStakedReward is the base block reward for epos.
-	BaseStakedReward = numeric.NewDecFromBigInt(new(big.Int).Mul(
-		big.NewInt(18), big.NewInt(denominations.One),
-	))
-	// BlockRewardStakedCase is the baseline block reward in staked case -
-	totalTokens                  = numeric.NewDecFromBigInt(new(big.Int).Mul(big.NewInt(12600000000), big.NewInt(denominations.One)))
-	targetStakedPercentage       = numeric.MustNewDecFromStr("0.35")
-	dynamicAdjust                = numeric.MustNewDecFromStr("0.4")
-	errPayoutNotEqualBlockReward = errors.New("total payout not equal to blockreward")
-	noReward                     = big.NewInt(0)
-)
-
-func adjust(amount numeric.Dec) numeric.Dec {
-	return amount.MulTruncate(
-		numeric.NewDecFromBigInt(big.NewInt(denominations.One)),
-	)
-}
 
 func ballotResultBeaconchain(
 	bc engine.ChainReader, header *block.Header,
 ) (shard.SlotList, shard.SlotList, shard.SlotList, error) {
 	return availability.BallotResult(bc, header, shard.BeaconChainShardID)
-}
-
-func whatPercentStakedNow(
-	beaconchain engine.ChainReader,
-	timestamp int64,
-) (*numeric.Dec, error) {
-	stakedNow := numeric.ZeroDec()
-	// Only active validators' stake is counted in stake ratio because only their stake is under slashing risk
-	active, err := beaconchain.ReadActiveValidatorList()
-	if err != nil {
-		return nil, err
-	}
-
-	soFarDoledOut, err := beaconchain.ReadBlockRewardAccumulator(
-		beaconchain.CurrentHeader().Number().Uint64(),
-	)
-
-	if err != nil {
-		return nil, err
-	}
-
-	dole := numeric.NewDecFromBigInt(soFarDoledOut)
-
-	for i := range active {
-		wrapper, err := beaconchain.ReadValidatorInformation(active[i])
-		if err != nil {
-			return nil, err
-		}
-		stakedNow = stakedNow.Add(
-			numeric.NewDecFromBigInt(wrapper.TotalDelegation()),
-		)
-	}
-	percentage := stakedNow.Quo(totalTokens.Mul(
-		reward.PercentageForTimeStamp(timestamp),
-	).Add(dole))
-	utils.Logger().Info().
-		Str("so-far-doled-out", dole.String()).
-		Str("staked-percentage", percentage.String()).
-		Str("currently-staked", stakedNow.String()).
-		Msg("Computed how much staked right now")
-	return &percentage, nil
 }
 
 // AccumulateRewards credits the coinbase of the given block with the mining
@@ -101,29 +39,26 @@ func AccumulateRewards(
 
 	if blockNum == 0 {
 		// genesis block has no parent to reward.
-		return noReward, nil
+		return network.NoReward, nil
 	}
 
 	if bc.Config().IsStaking(header.Epoch()) &&
 		bc.CurrentHeader().ShardID() != shard.BeaconChainShardID {
-		return noReward, nil
+		return network.NoReward, nil
 	}
 
 	//// After staking
 	if bc.Config().IsStaking(header.Epoch()) &&
 		bc.CurrentHeader().ShardID() == shard.BeaconChainShardID {
 
-		defaultReward := BaseStakedReward
+		defaultReward := network.BaseStakedReward
 
 		// TODO Use cached result in off-chain db instead of full computation
-		percentageStaked, err := whatPercentStakedNow(beaconChain, header.Time().Int64())
+		_, percentageStaked, err := network.WhatPercentStakedNow(beaconChain, header.Time().Int64())
 		if err != nil {
-			return noReward, err
+			return network.NoReward, err
 		}
-		howMuchOff := targetStakedPercentage.Sub(*percentageStaked)
-		adjustBy := adjust(
-			howMuchOff.MulTruncate(numeric.NewDec(100)).Mul(dynamicAdjust),
-		)
+		howMuchOff, adjustBy := network.Adjustment(*percentageStaked)
 		defaultReward = defaultReward.Add(adjustBy)
 		utils.Logger().Info().
 			Str("percentage-token-staked", percentageStaked.String()).
@@ -134,7 +69,7 @@ func AccumulateRewards(
 		// If too much is staked, then possible to have negative reward,
 		// not an error, just a possible economic situation, hence we return
 		if defaultReward.IsNegative() {
-			return noReward, nil
+			return network.NoReward, nil
 		}
 
 		newRewards := big.NewInt(0)
@@ -142,12 +77,12 @@ func AccumulateRewards(
 		// Take care of my own beacon chain committee, _ is missing, for slashing
 		members, payable, _, err := ballotResultBeaconchain(beaconChain, header)
 		if err != nil {
-			return noReward, err
+			return network.NoReward, err
 		}
 
 		votingPower, err := votepower.Compute(members)
 		if err != nil {
-			return noReward, err
+			return network.NoReward, err
 		}
 
 		for beaconMember := range payable {
@@ -157,7 +92,7 @@ func AccumulateRewards(
 			if !voter.IsHarmonyNode {
 				snapshot, err := bc.ReadValidatorSnapshot(voter.EarningAccount)
 				if err != nil {
-					return noReward, err
+					return network.NoReward, err
 				}
 				due := defaultReward.Mul(
 					voter.EffectivePercent.Quo(votepower.StakersShare),
@@ -172,7 +107,7 @@ func AccumulateRewards(
 			crossLinks := types.CrossLinks{}
 			err := rlp.DecodeBytes(cxLinks, &crossLinks)
 			if err != nil {
-				return noReward, err
+				return network.NoReward, err
 			}
 
 			type slotPayable struct {
@@ -193,7 +128,7 @@ func AccumulateRewards(
 				shardState, err := bc.ReadShardState(cxLink.Epoch())
 
 				if err != nil {
-					return noReward, err
+					return network.NoReward, err
 				}
 
 				subComm := shardState.FindCommitteeByID(cxLink.ShardID())
@@ -201,12 +136,12 @@ func AccumulateRewards(
 				payableSigners, _, err := availability.BlockSigners(cxLink.Bitmap(), subComm)
 
 				if err != nil {
-					return noReward, err
+					return network.NoReward, err
 				}
 
 				votingPower, err := votepower.Compute(payableSigners)
 				if err != nil {
-					return noReward, err
+					return network.NoReward, err
 				}
 				for j := range payableSigners {
 					voter := votingPower.Voters[payableSigners[j].BlsPublicKey]
@@ -250,7 +185,7 @@ func AccumulateRewards(
 				for payThem := range resultsHandle[bucket] {
 					snapshot, err := bc.ReadValidatorSnapshot(resultsHandle[bucket][payThem].payee)
 					if err != nil {
-						return noReward, err
+						return network.NoReward, err
 					}
 					due := resultsHandle[bucket][payThem].payout.TruncateInt()
 					newRewards = new(big.Int).Add(newRewards, due)
@@ -260,7 +195,7 @@ func AccumulateRewards(
 
 			return newRewards, nil
 		}
-		return noReward, nil
+		return network.NoReward, nil
 	}
 
 	//// Before staking
@@ -274,17 +209,17 @@ func AccumulateRewards(
 	if parentHeader.Number().Cmp(common.Big0) == 0 {
 		// Parent is an epoch block,
 		// which is not signed in the usual manner therefore rewards nothing.
-		return noReward, nil
+		return network.NoReward, nil
 	}
 
 	_, signers, _, err := availability.BallotResult(bc, header, header.ShardID())
 
 	if err != nil {
-		return noReward, err
+		return network.NoReward, err
 	}
 
 	totalAmount := rewarder.Award(
-		BlockReward, signers, func(receipient common.Address, amount *big.Int) {
+		network.BlockReward, signers, func(receipient common.Address, amount *big.Int) {
 			payable = append(payable, struct {
 				string
 				common.Address
@@ -294,13 +229,13 @@ func AccumulateRewards(
 		},
 	)
 
-	if totalAmount.Cmp(BlockReward) != 0 {
+	if totalAmount.Cmp(network.BlockReward) != 0 {
 		utils.Logger().Error().
-			Int64("block-reward", BlockReward.Int64()).
+			Int64("block-reward", network.BlockReward.Int64()).
 			Int64("total-amount-paid-out", totalAmount.Int64()).
 			Msg("Total paid out was not equal to block-reward")
 		return nil, errors.Wrapf(
-			errPayoutNotEqualBlockReward, "payout "+totalAmount.String(),
+			network.ErrPayoutNotEqualBlockReward, "payout "+totalAmount.String(),
 		)
 	}
