+++ staking/network/values.go
@@ -1,0 +1,9 @@
+package network
+
+import "math/big"
+
+// Reward ..
+type Reward struct {
+	AccumulatorSnapshot     *big.Int `json:"block-reward-accumulator"`
+	CurrentStakedPercentage *big.Int `json:"block-reward-accumulator"`
+}
